requirements.txt
# Core dependencies
google-generativeai>=0.3.0
speechrecognition>=3.10.0
pyttsx3>=2.90
wikipedia>=1.4.0
pywhatkit>=5.4
pyautogui>=0.9.54
gtts>=2.3.0
pyjokes>=0.6.0
pyaudio>=0.2.11
requests>=2.28.0
python-dotenv>=1.0.0
psutil>=5.9.0
pycaw>=20220416
screen-brightness-control>=0.21.0

# Enhanced features dependencies
qrcode>=7.4.2
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
schedule>=1.2.0
